+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:chatbot/chatbot.dart';
+import 'package:chatbot_launcher/chatbot.dart';
 
 void main() => runApp(const MyApp());
 
@@ -9,12 +9,12 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Chatbot Demo',
+      title: 'chatbot Demo',
       home: Scaffold(
-        appBar: AppBar(title: const Text("Chatbot Example")),
+        appBar: AppBar(title: const Text("chatbot Example")),
         body: const Center(child: Text("Home Screen")),
-        floatingActionButton: ChatbotLauncher(
-          config: ChatbotConfig(
+        floatingActionButton: chatbotLauncher(
+          config: chatbotConfig(
             fabColor: Colors.purple,
             headerTitle: "Hi! Need Help?",
             headerSubtitle: "Ask us anything — we’re here to help.",