+++ lib/src/utils/chatbot_assets.dart
@@ -1,4 +1,4 @@
-class ChatbotAssets {
+class chatbotAssets {
   // 🔹 Default assets
   static const String defaultBackground = 'assets/images/baground.jpg';
   static const String defaultLogo = 'assets/images/logo.png';